benchmarks/requirements.txt
requests>=2.28.0
pandas>=1.5.0
ruamel.yaml>=0.17.0
tqdm>=4.64.0
datasets>=2.14.0
